[Refactor] Remove useless analyze method and mini code format in RowDelimiter (#13060)
