+++ fe/fe-core/src/main/java/com/starrocks/analysis/BaseTableRef.java
@@ -19,8 +19,6 @@ package com.starrocks.analysis;
 
 import com.starrocks.catalog.Table;
 import com.starrocks.common.AnalysisException;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 /**
  * An actual table, such as OLAP table or a MySQL table.
