+++ fe/fe-core/src/main/java/com/starrocks/analysis/Delimiter.java
@@ -3,7 +3,7 @@
 package com.starrocks.analysis;
 
 import com.google.common.base.Strings;
-import com.starrocks.common.AnalysisException;
+import com.starrocks.sql.analyzer.SemanticException;
 
 import java.io.StringWriter;
 
@@ -26,23 +26,23 @@ public class Delimiter {
         return (byte) HEX_STRING.indexOf(c);
     }
 
-    public static String convertDelimiter(String originStr) throws AnalysisException {
+    public static String convertDelimiter(String originStr) {
         if (Strings.isNullOrEmpty(originStr)) {
-            throw new AnalysisException("Delimiter cannot be empty or null");
+            throw new SemanticException("Delimiter cannot be empty or null");
         }
 
         if (originStr.toUpperCase().startsWith("\\X") || originStr.toUpperCase().startsWith("0X")) {
             String hexStr = originStr.substring(2);
             // check hex str
             if (hexStr.isEmpty()) {
-                throw new AnalysisException("Invalid delimiter '" + originStr + ": empty hex string");
+                throw new SemanticException("Invalid delimiter '" + originStr + ": empty hex string");
             }
             if (hexStr.length() % 2 != 0) {
-                throw new AnalysisException("Invalid delimiter '" + originStr + ": hex length must be a even number");
+                throw new SemanticException("Invalid delimiter '" + originStr + ": hex length must be a even number");
             }
             for (char hexChar : hexStr.toUpperCase().toCharArray()) {
                 if (HEX_STRING.indexOf(hexChar) == -1) {
-                    throw new AnalysisException("Invalid delimiter '" + originStr + "': invalid hex format");
+                    throw new SemanticException("Invalid delimiter '" + originStr + "': invalid hex format");
                 }
             }
