+++ fe/fe-core/src/main/java/com/starrocks/analysis/ExprId.java
@@ -19,8 +19,6 @@ package com.starrocks.analysis;
 
 import com.starrocks.common.Id;
 import com.starrocks.common.IdGenerator;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public class ExprId extends Id<ExprId> {
