+++ fe/fe-core/src/main/java/com/starrocks/load/RoutineLoadDesc.java
@@ -21,14 +21,14 @@
 
 package com.starrocks.load;
 
-import com.starrocks.analysis.ColumnSeparator;
 import com.starrocks.analysis.Expr;
-import com.starrocks.analysis.RowDelimiter;
 import com.starrocks.analysis.SlotRef;
+import com.starrocks.sql.ast.ColumnSeparator;
 import com.starrocks.sql.ast.ImportColumnDesc;
 import com.starrocks.sql.ast.ImportColumnsStmt;
 import com.starrocks.sql.ast.ImportWhereStmt;
 import com.starrocks.sql.ast.PartitionNames;
+import com.starrocks.sql.ast.RowDelimiter;
 
 import java.util.ArrayList;
 import java.util.List;
