+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/RangePartitionDesc.java
@@ -132,7 +132,7 @@ public class RangePartitionDesc extends PartitionDesc {
                 if (column.getName().equalsIgnoreCase(colName)) {
                     if (!column.isKey() && column.getAggregationType() != AggregateType.NONE) {
                         throw new DdlException("The partition column could not be aggregated column"
-                                    + " and unique table's partition column must be key column");
+                                + " and unique table's partition column must be key column");
                     }
 
                     if (column.getType().isFloatingPointType() || column.getType().isComplexType()) {
